refactor: extract wait semantics to worker_wait.go
